# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FeedbackLab API",
    "description": "API для управления проектами, тикетами и документацией",
    "version": "0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://localhost:8080",
      "description": "Local HTTPS server"
    }
  ],
...
  "paths": {
...
    "/tickets": {
      "post": {
        "tags": ["Tickets"],
        "summary": "Создать тикет",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/": {
      "get": {
        "tags": ["Tickets"],
        "summary": "Получить все тикеты",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ticket"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{id}": {
      "get": {
        "tags": ["Tickets"],
        "summary": "Получить тикет по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID тикета"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Tickets"],
        "summary": "Обновить тикет",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID тикета"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Tickets"],
        "summary": "Удалить тикет",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID тикета"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/ticket_chats/": {
      "post": {
        "tags": ["TicketChats"],
        "summary": "Создать сообщение тикета",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketChatDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketChat"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/ticket_chats/{id}": {
      "get": {
        "tags": ["TicketChats"],
        "summary": "Получить сообщение по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID сообщения"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketChat"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["TicketChats"],
        "summary": "Обновить сообщение",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID сообщения"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketChatDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketChat"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["TicketChats"],
        "summary": "Удалить сообщение",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID сообщения"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/ticket_chats/ticket/{ticket_id}": {
      "get": {
        "tags": ["TicketChats"],
        "summary": "Получить все сообщения тикета",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID тикета"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketChat"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Ticket": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "project_id": {
            "type": "integer"
          },
          "module_id": {
            "type": "integer",
            "nullable": true
          },
          "contract_id": {
            "type": "integer"
          },
          "created_by": {
            "type": "string"
          },
          "assigned_to": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["open", "in_progress", "resolved", "closed"]
          },
          "gitlab_issue_url": {
            "type": "string",
            "nullable": true
          },
          "mattermost_thread_url": {
            "type": "string",
            "nullable": true
          },
          "date_created": {
            "type": "string",
            "format": "date-time"
          },
          "date_updated": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["project_id", "contract_id", "created_by", "title", "message"]
      },
      "CreateTicketDTO": {
        "type": "object",
        "properties": {
          "project_id": {
            "type": "integer"
          },
          "module_id": {
            "type": "integer"
          },
          "contract_id": {
            "type": "integer"
          },
          "created_by": {
            "type": "string"
          },
          "assigned_to": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["open", "in_progress", "resolved", "closed"]
          },
          "gitlab_issue_url": {
            "type": "string"
          },
          "mattermost_thread_url": {
            "type": "string"
          }
        },
        "required": ["project_id", "contract_id", "created_by", "title", "message"]
      },
      "UpdateTicketDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["open", "in_progress", "resolved", "closed"]
          },
          "assigned_to": {
            "type": "string"
          },
          "gitlab_issue_url": {
            "type": "string"
          },
          "mattermost_thread_url": {
            "type": "string"
          }
        },
        "required": ["title", "message", "status"]
      },
      "TicketChat": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "ticket_id": {
            "type": "integer"
          },
          "sender_id": {
            "type": "string"
          },
          "sender_role": {
            "type": "string",
            "enum": ["customer", "manager", "developer", "tester"]
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "message_type": {
            "type": "string",
            "enum": ["text", "file", "system"]
          },
          "mattermost_message_id": {
            "type": "string",
            "nullable": true
          },
          "date_created": {
            "type": "string",
            "format": "date-time"
          },
          "date_updated": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["ticket_id", "sender_id", "sender_role", "message"]
      },
      "CreateTicketChatDTO": {
        "type": "object",
        "properties": {
          "ticket_id": {
            "type": "integer"
          },
          "sender_id": {
            "type": "string"
          },
          "sender_role": {
            "type": "string",
            "enum": ["customer", "manager", "developer", "tester"]
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "message_type": {
            "type": "string",
            "enum": ["text", "file", "system"]
          },
          "mattermost_message_id": {
            "type": "string"
          }
        },
        "required": ["ticket_id", "sender_id", "sender_role", "message"]
      },
      "UpdateTicketChatDTO": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1
          },
          "message_type": {
            "type": "string",
            "enum": ["text", "file", "system"]
          }
        }
      },
...
    }
  }
}